components/resume_manager/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
gunicorn==20.1.0
google-cloud-storage==2.10.0
